+++ src/layman/map/filesystem/input_file.py
@@ -149,3 +149,8 @@ def post_map(username, mapname, description, title):
 
 
 patch_map = post_map
+
+
+def get_metadata_comparison(username, publication_name):
+    pass
+